Moves some functions to tools
